[CSS] Split the h2 and h2 a rules
Split the h2 and h2 a rules. The h2 now styles the area and the h2 a
handles the anchor.

Now the text doesn't write over itself when it wraps. At least when it
overlays itself now, it has a transparent background and you can't tell.

Same thing? :P

Also removed the height property from the ::after that displays "1up"
as the height was superfluous.